+++ apps/cowswap-frontend/src/modules/advancedOrders/hooks/useComposableCowContract.ts
@@ -1,9 +1,22 @@
-import { ComposableCoW, ComposableCoWAbi } from '@cowprotocol/cowswap-abis'
+import { useMemo } from 'react'
 
-import { useContract, UseContractResult } from 'common/hooks/useContract'
+import { ComposableCoWAbi } from '@cowprotocol/cowswap-abis'
+import { useWalletInfo } from '@cowprotocol/wallet'
+
+import { UseContractResult } from 'common/hooks/useContract'
 
 import { COMPOSABLE_COW_ADDRESS } from '../const'
 
-export function useComposableCowContract(): UseContractResult<ComposableCoW> {
-  return useContract<ComposableCoW>(COMPOSABLE_COW_ADDRESS, ComposableCoWAbi, true)
+export type ComposableCowContractData = UseContractResult<typeof ComposableCoWAbi>
+export function useComposableCowContractData(): ComposableCowContractData {
+  const { chainId } = useWalletInfo()
+
+  return useMemo(
+    () => ({
+      abi: ComposableCoWAbi,
+      address: COMPOSABLE_COW_ADDRESS[chainId],
+      chainId,
+    }),
+    [chainId],
+  )
 }