+++ apps/cowswap-frontend/src/modules/ordersTable/pure/OrdersTabs/OrdersTabs.pure.tsx
@@ -1,7 +1,7 @@
 import { ReactNode, ChangeEvent } from 'react'
 
-import alertCircle from '@cowprotocol/assets/cow-swap/alert-circle.svg'
-import orderPresignaturePending from '@cowprotocol/assets/cow-swap/order-presignature-pending.svg'
+import iconInformation from '@cowprotocol/assets/cow-swap/filled-info-circle.svg'
+import iconOrderPresignaturePending from '@cowprotocol/assets/cow-swap/order-presignature-pending.svg'
 import { useWalletInfo } from '@cowprotocol/wallet'
 
 import { t } from '@lingui/core/macro'
@@ -65,8 +65,8 @@ export function OrdersTabs({ tabs }: OrdersTabsProps): ReactNode {
               $isDisabled={!account}
               to={buildOrdersTableUrl({ tabId: tab.id, pageNumber: 1 })}
             >
-              {isUnfillable && <SVG src={alertCircle} description={t`warning`} />}
-              {isSigning && <SVG src={orderPresignaturePending} description={t`signing`} />}
+              {isUnfillable && <SVG src={iconInformation} description={t`warning`} />}
+              {isSigning && <SVG src={iconOrderPresignaturePending} description={t`signing`} />}
               {i18n._(tab.title)} {account && <span>({tab.count})</span>}
             </styledEl.TabButton>
           )